pyproject.toml
[project]
name = "ingestlab"
version = "6.2.2"
description = "Document ingestion sandbox for Azure Document Intelligence with custom chunking."
requires-python = ">=3.11,<3.12"

# Core dependencies - web server and document processing
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "azure-ai-documentintelligence",
    "requests",
    "python-dotenv",
    "python-multipart",
    "pypdf",
    "openai",
    # PolicyAsCode extractors and figure processing
    "brd-to-opa-pipeline @ git+https://github.com/kyndryl-agentic-ai/PolicyAsCode.git@feature/chunking-visualizer-integration",
    # Required for figure processing (FigureProcessor)
    "pymupdf>=1.26.7",
    "opencv-python>=4.13.0.90",
]

[project.optional-dependencies]
# For running SAM3 model server locally (GPU recommended)
# Not needed if using remote SAM3 endpoint (RunPod, etc.)
sam3-local = [
    "torch>=2.10.0",
    "transformers>=5.0.0rc1",
    "huggingface-hub>=0.36.0",
    "torchvision>=0.25.0",
]

# For deploying SAM3 to Modal serverless (GPU)
# Deploy from PolicyAsCode: uv run modal deploy .venv/lib/python3.11/site-packages/src/services/sam3_modal_app.py
modal = [
    "modal>=0.64.0",
]

# Full install with all optional features
full = [
    "ingestlab[sam3-local]",
    "ingestlab[modal]",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []
